requirements_duckdb.txt
duckdb>=1.1.0
boto3>=1.34.0
pandas>=2.2.0
yfinance>=0.2.54
pyarrow>=16.0.0
polars>=1.5.0
pykrx>=1.2.3
finance-datareader>=0.9.101
python-dotenv>=1.0.1
